MAINT: PR 28 revisions
* move the new ufunc tests into their
own separate testing module, based on
reviewer feedback
